web/requirements.txt
Flask
torch
flask_restful
bcrypt
requests
ultralytics
numpy
pillow
mlflow
Flask==3.0.3
Flask-RESTful==0.3.9
torch==1.10.0
torchvision==0.11.1
torchaudio==0.10.0